Set up project
Set up an Eclipse project with a simple Hello World, got it to build
using MinGW gcc on Windows 10. So far the project is very
Windows-centric, but the point of Eclipse is so it doesn't have to be.
I'll probably work next on getting a single project file to work on both
Windows and Linux.